% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_malaria.R
\name{theme_malaria}
\alias{theme_malaria}
\title{Theme for ggplots}
\usage{
theme_malaria(base_size = 14)
}
\arguments{
\item{base_size}{The base font size}
}
\value{
A ggplot compatible theme call
}
\description{
Add a theme to a ggplot object
}

